Uncomment older VC skip
We have been unable to build this feedstock against older VC versions
since it has existed. In fact it is still failing currently. So go ahead
and uncomment this skip to ignore the builds with older VC versions.
These will be dropped soon anyways (with the EOL of Python 2 fast
approaching).